# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "REST API для управления проектами и задачами с системой авторизации, назначением исполнителей и поддержкой файловых вложений.",
        "contact": {
            "email": "support@task-manager.ru"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://task-manager.local",
            "description": "Локальный сервер разработки"
        },
        {
            "url": "https://dev.xxxtask-manager.ru",
            "description": "DEV сервер"
        },
        {
            "url": "https://prod.xxxtask-manager.ru",
            "description": "PROD сервер"
        }
    ],
    "paths": {
...
        "/api/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу по ID",
                "operationId": "32c40e2d1129b80c1dcabd538df63a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные задачи",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "operationId": "c36c959b634c99e1e02ad2f3f2c36a2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Обновленная задача"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Новое описание",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "planned",
                                            "in_progress",
                                            "done"
                                        ],
                                        "example": "in_progress"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31",
                                        "nullable": true
                                    },
                                    "assignee_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "attachments": {
                                        "description": "Массив файлов (заменит существующие)",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "d81cc7975ea62176b4ef8f74ed95168d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Задача успешно удалена"
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Задача не найдена"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "TaskResource": {
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "project_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "title": {
                                "type": "string",
                                "example": "Моя задача"
                            },
                            "description": {
                                "type": "string",
                                "example": "Описание задачи",
                                "nullable": true
                            },
                            "status": {
                                "type": "string",
                                "example": "planned"
                            },
                            "due_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2025-12-31",
                                "nullable": true
                            },
                            "assignee": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Иван Иванов"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ivan@example.com"
                                    }
                                },
                                "type": "object",
                                "nullable": true
                            },
                            "attachments": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "document"
                                        },
                                        "file_name": {
                                            "type": "string",
                                            "example": "document.pdf"
                                        },
                                        "mime_type": {
                                            "type": "string",
                                            "example": "application/pdf"
                                        },
                                        "size": {
                                            "type": "integer",
                                            "example": 102400
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost/storage/1/document.pdf"
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен полученный после регистрации или входа",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}